Improved ifdef checks for Client and Server classes. Changed many functions to 'const'. Fixed up compiler warnings. MapVec3 + and - operators now return a value rather than reference to local variable. Changed Read and Write functions in several classes to simply take in a reference to a BitStream. This will allow more data to be sent in a single packet rather than sending a packet for several classes of data at once.
Added functions to convert a tile position to a world position using a tile-size scale.
BaseAction class now also has a static float 'm_simulationTime'. This will allow newly connected clients to simulate the game fast-forwarding.
ShootAction constructor no longer needs to know if the shot was a critical or not.
Game class is now a proper singleton. (I somehow forgot to make the GetInstance() function or the instance pointer static, which meant it was not previously working as a singleton.)

Server now stores connected client IPs in an std::map. This will later be used to hand re-connecting to the same position, etc.